+++ NEWS.md
@@ -2,9 +2,10 @@
 
 ## Changes compared to last release
 
+- Support for logprobs in `openai_chat()` and new `get_logprobs()` function (experimental!)
 - Bugfix in OpenAI metadata extraction
 - New `ollama_delete_model()` function
-- `list_models()` is now a verb supporting both `openai()`,`claude()` and `ollama()`. Other providers will be added
+- `list_models()` is now a verb supporting most providers.
 - Ellmer schemata supported in `.json_schema` arguements
 - New `send_ollama_batch()` function to make use of the fast parallel request features of Ollama. 
 - New batch functions for Azure Openai (thanks [Jia Zhang](https://github.com/JiaZhang42))